// src/types/tweetdeck.ts
interface ChirpAccount {
  state: ChirpAccountState;
  privateState: ChirpAccountPrivateState;
  managed: boolean;
}

interface ChirpAccountPrivateState {
  isPrivate: boolean;
  verified: boolean;
  updated: number;
  key: string;
}

interface ChirpAccountState {
  type: string;
  oauth_token: string;
  username: string;
  name: string;
  profileImageURL: string;
  userId: string;
  require_some_consent: boolean;
}

interface APIBounds {
  [k: string]: ApiBound;
}

interface ApiBound {
  lower: string;
  upper: number;
}

interface HighWaterMarkChirpSortIndex {
  value: number;
  type: string;
}

interface User {
  account: ChirpAccount;
  id: string;
  screenName: string;
  profileURL: string;
  profileImageURL: string;
  avatarSuffixRegex: unknown;
  name: string;
  emojifiedName: string;
  location: string;
  description: string;
  entities: SourceUserEntities;
  joinedDate: Date;
  friendsCount: number;
  listedCount: number;
  followersCount: number;
  statusesCount: number;
  url: string;
  isProtected: boolean;
  isVerified: boolean;
  isTranslator: boolean;
  isBadged: boolean;
  following: boolean;
  lang: null;
  profileColor: string;
  bannerUrl: string;
  bannerUrlSmall: string;
  _profileBannerURL: string;
}

interface SourceUserEntities {
  url: UrlEntity;
  description: UrlEntity;
}

interface UrlEntity {
  urls: TwitterUrl[];
}

interface TwitterUrl {
  url: string;
  expanded_url: string;
  display_url: string;
  indices: number[];
}

export interface TweetDeckChirp {
  account: ChirpAccount;
  id: string;
  sortIndex: TargetTweetSortIndex;
  user: User;
  sourceNoHTML: string;
  sourceUrl: string;
  prettyRetweetCount: boolean;
  prettyLikeCount: boolean;
  prettyReplyCount: boolean;
  withPrettyEngagements: boolean;
  inReplyToUserID: string;
  inReplyToScreenName: string;
  isFavorite: boolean;
  isRetweeted: boolean;
  retweetCount: number;
  likeCount: number;
  replyCount: number;
  inReplyToID: string | number | null;
  selfThreadId: boolean;
  text: string;
  entities: TweetEntities;
  isQuoteStatus: boolean;
  quotedTweet?: TweetDeckChirp;
  targetTweet?: TweetDeckChirp;
  retweet?: TweetDeckChirp;
  retweetedStatus?: TweetDeckChirp;
  quotedTweetMissing: boolean;
  _media: any[];
  lang: string;
  place: unknown;
  created: Date;
  conversationMuted: boolean;
  htmlText: string;
  chirpType?: string;
  _btd?: {
    chirpKey: string;
    columnKey: string;
  };
  renderInMediaGallery(): string;
}

export interface TweetHashtag {
  indices: [number, number];
  text: string;
}

export interface TweetEntities {
  hashtags: Twitter.HashtagEntity[];
  urls: Twitter.UrlEntity[];
  user_mentions: Twitter.UserMentionEntity[];
  cashtags: any[];
  media: Twitter.MediaEntity[];
  convertedToUTF16: boolean;
}

interface TargetTweetSortIndex {
  value: string;
  type: string;
}